# backend/src/main/resources/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
...
paths:
...
  /eventUsers/{id}/event/{propertyId}:
    get:
      tags:
        - event-user-property-reference-controller
      description: get-event-by-eventuser-Id
      operationId: followPropertyReference-eventuser-get
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: propertyId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelEvent'
        '404':
          description: Not Found
    delete:
      tags:
        - event-user-property-reference-controller
      description: delete-event-by-eventuser-Id
      operationId: deletePropertyReferenceId-eventuser-delete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: propertyId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: No Content
        '404':
          description: Not Found
...
  /events:
    get:
      tags:
        - event-entity-controller
      description: get-event
      operationId: getCollectionResource-event-get_1
      responses:
        '200':
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelEntityModelEvent'
            application/x-spring-data-compact+json:
              schema:
                $ref: '#/components/schemas/CollectionModelEntityModelEvent'
            text/uri-list:
              schema:
                type: string
    post:
      tags:
        - event-entity-controller
      description: create-event
      operationId: postCollectionResource-event-post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventRequestBody'
        required: true
      responses:
        '201':
          description: Created
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelEvent'
...
  /events/{id}:
    get:
      tags:
        - event-entity-controller
      description: get-event
      operationId: getItemResource-event-get
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelEvent'
        '404':
          description: Not Found
    put:
      tags:
        - event-entity-controller
      description: update-event
      operationId: putItemResource-event-put
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventRequestBody'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelEvent'
        '201':
          description: Created
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelEvent'
        '204':
          description: No Content
    delete:
      tags:
        - event-entity-controller
      description: delete-event
      operationId: deleteItemResource-event-delete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: No Content
        '404':
          description: Not Found
    patch:
      tags:
        - event-entity-controller
      description: patch-event
      operationId: patchItemResource-event-patch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventRequestBody'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelEvent'
        '204':
          description: No Content
...
  /rest/events:
    post:
      tags:
        - Event
      summary: Создать событие
      operationId: create
      parameters:
        - name: event
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/Event'
      responses:
        '200':
          description: Успешно
          content:
            application/hal+json:
              schema:
                type: integer
                format: int64
        '500':
          description: Internal Server Error
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/Exception'
...
  /rest/locations/{id}/events:
    get:
      tags:
        - Location
      summary: Получить список всех событий локации
      operationId: getEvents
      parameters:
        - name: id
          in: path
          description: Идентификатор локации
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Успешно
          content:
            application/hal+json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
        '500':
          description: Internal Server Error
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/Exception'
  /rest/events/{id}:
    get:
      tags:
        - Event
      summary: Получить событие
      operationId: getById_2
      parameters:
        - name: id
          in: path
          description: Идентификатор события
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Успешно
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/Event'
        '404':
          description: Событие не найдено
        '500':
          description: Internal Server Error
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/Exception'
  /rest/events/{id}/users:
    get:
      tags:
        - Event
      summary: Получить список участников события
      operationId: getUsers
      parameters:
        - name: id
          in: path
          description: Идентификатор события
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Успешно
          content:
            application/hal+json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '500':
          description: Internal Server Error
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/Exception'
  /rest/events/active:
    get:
      tags:
        - Event
      summary: Получить список активных событий
      operationId: getActive
      responses:
        '200':
          description: Успешно
          content:
            application/hal+json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
        '500':
          description: Internal Server Error
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/Exception'
components:
  schemas:
    Exception:
      type: object
      properties:
        message:
          type: string
...
    Links:
      type: object
      additionalProperties:
        $ref: '#/components/schemas/Link'
...
    EntityModelEvent:
      type: object
      properties:
        creationDate:
          type: string
          format: date-time
        startDate:
          type: string
          format: date-time
        duration:
          type: integer
          format: int64
        totalMembers:
          type: integer
          format: int32
        countMembers:
          type: integer
          format: int32
        _links:
          $ref: '#/components/schemas/Links'
    Event:
      type: object
      properties:
        id:
          type: integer
          format: int64
        creationDate:
          type: string
          format: date-time
        startDate:
          type: string
          format: date-time
        duration:
          type: integer
          format: int64
        totalMembers:
          type: integer
          format: int32
        countMembers:
          type: integer
          format: int32
        type:
          $ref: '#/components/schemas/EventType'
        location:
          $ref: '#/components/schemas/Location'
    EventType:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
    Location:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        latitude:
          type: number
          format: float
        longitude:
          type: number
          format: float
        address:
          type: string
    CollectionModelEntityModelEvent:
      type: object
      properties:
        _embedded:
          type: object
          properties:
            events:
              type: array
              items:
                $ref: '#/components/schemas/EntityModelEvent'
        _links:
          $ref: '#/components/schemas/Links'
...
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        fullName:
          type: string
        phoneNumber:
          type: string
        avatarPath:
          type: string
...
    EventRequestBody:
      type: object
      properties:
        id:
          type: integer
          format: int64
        creationDate:
          type: string
          format: date-time
        startDate:
          type: string
          format: date-time
        duration:
          type: integer
          format: int64
        totalMembers:
          type: integer
          format: int32
        countMembers:
          type: integer
          format: int32
        type:
          type: string
        location:
          type: string
...
    Link:
      type: object
      properties:
        href:
          type: string
        hreflang:
          type: string
        title:
          type: string
        type:
          type: string
        deprecation:
          type: string
        profile:
          type: string
        name:
          type: string
        templated:
          type: boolean